likelihood maximization keeps track of all combinations and returns them.
